requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0
Pillow>=10.0.0
redis>=5.0.0
